backend_api/requirements.txt
fastapi
uvicorn
pydantic
chromadb
sentence-transformers
yt-dlp
google-generativeai
youtube-transcript-api
python-dotenv
